refactor: 중복된 restoreAuth 제거
